增加readLoop，从mongo中持续读取数据并返回给客户端

+++ src/main/java/com/shahuwang/jmgo/ReplyOp.java
@@ -5,10 +5,10 @@ package com.shahuwang.jmgo;
  */
 public class ReplyOp {
     private int flags;
-    private int cursortId;
+    private long cursortId;
     private int firstDoc;
     private int replyDocs;
-    public ReplyOp(int flags, int cursortId, int firstDoc, int replyDocs){
+    public ReplyOp(int flags, long cursortId, int firstDoc, int replyDocs){
         this.firstDoc = firstDoc;
         this.flags = flags;
         this.cursortId = cursortId;
